# NUI-867 - Image post-processing: handle workers that cannot generate the rendition format