--- a/lib/postprocessing/image.js
+++ b/lib/postprocessing/image.js
@@ -32,14 +32,20 @@ const SUPPORTED_OUTPUT_FMT = {
 };
 
 // list of supported input formats (intermediate rendition)
-// in imagemagick format https://imagemagick.org/script/formats.php
-const SUPPORTED_INPUT_FMT = new Set([
-    "PNG",
-    "JPEG",
-    "TIFF",
-    "GIF",
-    "BMP"
-]);
+// maps imagemagick format ('PNG') https://imagemagick.org/script/formats.php to our fmt ('png')
+const SUPPORTED_INPUT_TYPES = {
+    "PNG": "png",
+    "JPEG": "jpeg",
+    "TIFF": "tif",
+    "GIF": "gif",
+    "BMP": "bmp"
+};
+const SUPPORTED_INPUT_FORMAT_IMAGEMAGICK = Object.keys(SUPPORTED_INPUT_TYPES);
+const SUPPORTED_INPUT_FMT = Object.values(SUPPORTED_INPUT_TYPES);
+
+const DEFAULT_WORKER_SUPPORTED_FMT = [
+    "png"
+];
 
 // list of post processing instructions
 //
@@ -70,7 +76,7 @@ function isSupportedOutputFormat(instructions) {
 }
 
 function isSupportedInputFormat(metadata) {
-    return SUPPORTED_INPUT_FMT.has(metadata.format);
+    return SUPPORTED_INPUT_FORMAT_IMAGEMAGICK.includes(metadata.format);
 }
 
 function hasExtraInstructions(instructions) {
@@ -142,9 +148,41 @@ function hasJpegQuality(metadata, instructions) {
     return true;
 }
 
+/**
+ * Returns updated rendition instructions for the worker callback processing if needed
+ * or undefined if no changes are required.
+ */
+async function prepareImagePostProcess(rendition, options) {
+    if (rendition.instructions.fmt && isSupportedOutputFormat(rendition.instructions)) {
+
+        const supportedRenditionFormats = options.supportedRenditionFormats || DEFAULT_WORKER_SUPPORTED_FMT;
+        // in case the worker cannot generate the final rendition format itself, we have to do it via post processing
+        if (!supportedRenditionFormats.includes(rendition.instructions.fmt)) {
+            // ...and tell the worker to generate a different intermediate format it can do
+            // find first format that is supported (worker defines priority order)
+            const newFormat = supportedRenditionFormats.find(fmt => SUPPORTED_INPUT_FMT.includes(fmt));
+            if (!newFormat) {
+                throw new GenericError(`Worker does not support generating a compatible image format. Requested: ${SUPPORTED_INPUT_FMT} Supported by worker: ${supportedRenditionFormats}`);
+            }
+
+            console.log(`post-processing: adjusted fmt for worker callback: '${rendition.instructions.fmt}' -> '${newFormat}'`);
+
+            rendition._forceImagePostProcess = true;
+
+            return Object.assign({}, rendition.instructions, { fmt: newFormat});
+        }
+    }
+
+    // return undefined if no instruction change needed
+}
+
+/**
+ * Returns true if the (intermediate) rendition needs image post processing.
+ */
 async function needsImagePostProcess(rendition) {
     const instructions = rendition.instructions;
 
+    // NOTE: already checked in prepareImagePostProcess() above, no need to check again?
     // verify if the requested format is one we can output from post-processing
     if (!isSupportedOutputFormat(instructions)) {
         console.log(`skipping post-processing because output fmt '${instructions.fmt}' is not supported`);
@@ -154,7 +192,12 @@ async function needsImagePostProcess(rendition) {
     await logImageMagickVersion();
 
     // inspect intermediate rendition
-    const metadata = rendition.intermediateMetadata = await getImageMetadata(rendition.path);
+    const metadata = rendition._intermediateMetadata = await getImageMetadata(rendition.path);
+
+    // if preparePostProcess() already determined we need post-processing
+    if (rendition._forceImagePostProcess) {
+        return true;
+    }
 
     // check if it can read the intermediate rendition
     if (!isSupportedInputFormat(metadata)) {
@@ -210,18 +253,18 @@ function getImageSize(metadata) {
         return new Dimensions(metadata.pageGeometry);
     }
 }
-async function applyWatemark(img, rendition, directories) {
+async function applyWatemark(img, intermediateRendition, rendition, directories) {
     const instructions = rendition.instructions;
     const targetBoundingBox = new Dimensions(instructions.width, instructions.height);
 
     const watermarkScale = instructions.watermark.scale || 1.0;
     // determine image dimensions
-    const imageSize = getImageSize(rendition.intermediateMetadata);
+    const imageSize = getImageSize(intermediateRendition._intermediateMetadata);
 
     const watermarkSource = instructions.watermark.image;
     // download watermark
     if(!watermarkSource) {
-        throw new GenericError(`post-processing: Missing watermark.image in instructions : ${JSON.stringify(rendition.instructions)}`);
+        throw new GenericError(`post-processing: Missing watermark.image in instructions : ${JSON.stringify(instructions)}`);
     }
     const asset = await getAsset(watermarkSource, directories.in, WATERMARK_NAME);
     // TODO: infuture should we enforce that watermark is a png ?
@@ -235,7 +278,7 @@ async function applyWatemark(img, rendition, directories) {
     img.draw([`gravity Center image Over 0,0 ${watermarkTargetSize.width},${watermarkTargetSize.height} "${asset.path}"`]);
 }
 
-async function applyOperations(img, rendition, directories) {
+async function applyOperations(img, intermediateRendition, rendition, directories) {
     const instructions = rendition.instructions;
     // handle exif orientation
     // MUST happen before img.resize() for orientation 5 to 8 (otherwise width & height must be swapped)
@@ -257,7 +300,7 @@ async function applyOperations(img, rendition, directories) {
     }
 
     if (typeof instructions.watermark === 'object'){
-        await applyWatemark(img, rendition, directories);
+        await applyWatemark(img, intermediateRendition, rendition, directories);
     }
 }
 
@@ -346,18 +389,24 @@ async function write(img, outfile) {
     await img.write(outfile);
 }
 
-// exported function
-async function imagePostProcess(rendition, outfile, directories) {
-    const img = await read(rendition.path);
+/**
+ * Runs image post processing on the (intermediate) rendition, to create the final rendition.
+ * @param {*} intermediateRendition existing intermediate rendition created by the worker callback
+ * @param {*} rendition final rendition to write to, including instructions
+ * @param {*} directories working directories of the current activation
+ */
+async function imagePostProcess(intermediateRendition, rendition, directories) {
+    const img = await read(intermediateRendition.path);
 
-    await applyOperations(img, rendition, directories);
+    await applyOperations(img, intermediateRendition, rendition, directories);
 
     await applyOutputProperties(img, rendition.instructions);
 
-    await write(img, outfile);
+    await write(img, rendition.path);
 }
 
 module.exports = {
+    prepareImagePostProcess,
     needsImagePostProcess,
     imagePostProcess
 };