--- a/lib/postprocessing/image.js
+++ b/lib/postprocessing/image.js
@@ -33,11 +33,12 @@ const SUPPORTED_OUTPUT_FMT = {
 
 // list of supported input formats (intermediate rendition)
 // maps imagemagick format ('PNG') https://imagemagick.org/script/formats.php to our fmt ('png')
+// Note: order is important and defines priority for intermediate format - first entry will be most preferred
 const SUPPORTED_INPUT_TYPES = {
-    "PNG": "png",
-    "JPEG": "jpeg",
-    "TIFF": "tif",
-    "GIF": "gif",
+    "TIFF": "tif",  // most preferred format as it allows to preserve many features
+    "PNG": "png",   // lossless over
+    "JPEG": "jpeg", // ..lossy
+    "GIF": "gif",   // (should never get here in practice)
     "BMP": "bmp"
 };
 const SUPPORTED_INPUT_FORMAT_IMAGEMAGICK = Object.keys(SUPPORTED_INPUT_TYPES);
@@ -46,6 +47,48 @@ const SUPPORTED_INPUT_FMT = Object.values(SUPPORTED_INPUT_TYPES);
 const DEFAULT_WORKER_SUPPORTED_FMT = [
     "png"
 ];
+
+// NOTE: this is a temporary solution until workers have been updated to set options.supportedRenditionFormats
+const WORKER_SUPPORTED_FMT = {
+    // https://git.corp.adobe.com/nui/worker-flite/blob/0e6dbe1e68ad9f226658bea3f7ba74c27a70073f/action/worker.js#L176
+    // https://git.corp.adobe.com/nui/core/blob/5e1dc52e950a38f93599f1095b07ae163c40dac2/lib/workers.js#L225
+    "worker-flite":     ["png", "jpg", "jpeg", "gif", "tif", "tiff"],
+
+    // https://git.corp.adobe.com/varya/imagecore_main/blob/9_9_linux/camera_raw/DAM_Raw_Converter/source/main.cpp#L203-L224
+    // https://git.corp.adobe.com/nui/core/blob/5e1dc52e950a38f93599f1095b07ae163c40dac2/lib/workers.js#L142
+    "worker-cameraraw": ["jpg", "jpeg", "tif", "tiff"],
+
+    // https://git.corp.adobe.com/nui/worker-pie/blob/925d73ad1abcb1eb7c6755eb544908250494bc11/action/worker.js#L58
+    // https://git.corp.adobe.com/nui/core/blob/5e1dc52e950a38f93599f1095b07ae163c40dac2/lib/workers.js#L180
+    "worker-pie":       ["png", "jpg", "jpeg", "gif", "tif", "tiff", "psd"],
+    "worker-pie-large": ["png", "jpg", "jpeg", "gif", "tif", "tiff", "psd"],
+
+    // https://git.corp.adobe.com/nui/worker-ffmpeg/blob/40ee89d248b03bb3a6c18f4b41288f9ef456a147/action/worker.js#L28
+    // https://git.corp.adobe.com/nui/core/blob/5e1dc52e950a38f93599f1095b07ae163c40dac2/lib/workers.js#L200
+    "worker-ffmpeg":    ["png", "jpg", "jpeg", "gif", "tif", "tiff"],
+
+    // https://git.corp.adobe.com/nui/worker-libreoffice/blob/master/action/worker.sh#L12
+    // https://git.corp.adobe.com/nui/core/blob/5e1dc52e950a38f93599f1095b07ae163c40dac2/lib/workers.js#L195
+    "worker-libreoffice": ["png"],
+
+    // https://git.corp.adobe.com/nui/worker-pdfrasterizer/blob/f939ab6f145ec698e0ccd37dd4a0b502a68c1fab/action/worker.sh#L110
+    // https://git.corp.adobe.com/nui/core/blob/5e1dc52e950a38f93599f1095b07ae163c40dac2/lib/workers.js#L213
+    "worker-pdfrasterizer": ["png"],
+
+    // ---------------------------------------------------------------
+    // below workers give us whatever is embedded as thumbnail format
+    // and it does not matter which fmt they are instructed to do (but the array must not be empty)
+
+    // https://git.corp.adobe.com/nui/worker-indesign/blob/6e97be7b0cd25d9eaf624d3a8200a1375abebad5/worker.js#L114
+    // per spec the embeded thumbnails in xmpGImg:format can only be JPEG (?)
+    //    https://github.com/adobe/xmp-docs/blob/master/XMPNamespaces/XMPDataTypes/Thumbnails.md
+    // https://git.corp.adobe.com/nui/core/blob/5e1dc52e950a38f93599f1095b07ae163c40dac2/lib/workers.js#L218
+    "worker-indesign": ["jpg"],
+
+    // https://git.corp.adobe.com/nui/worker-dcx/blob/605cd388ced64c8dfd789b40a51140ccc42be504/worker.js#L125
+    // https://git.corp.adobe.com/nui/core/blob/5e1dc52e950a38f93599f1095b07ae163c40dac2/lib/workers.js#L169
+    "worker-dcx": ["png"],
+};
 
 // list of post processing instructions
 //
@@ -155,14 +198,18 @@ function hasJpegQuality(metadata, instructions) {
 async function prepareImagePostProcess(rendition, options) {
     if (rendition.instructions.fmt && isSupportedOutputFormat(rendition.instructions)) {
 
-        const supportedRenditionFormats = options.supportedRenditionFormats || DEFAULT_WORKER_SUPPORTED_FMT;
+        const supportedFormats =
+            options.supportedRenditionFormats
+            || WORKER_SUPPORTED_FMT[process.env.__OW_ACTION_NAME]
+            || DEFAULT_WORKER_SUPPORTED_FMT;
+
         // in case the worker cannot generate the final rendition format itself, we have to do it via post processing
-        if (!supportedRenditionFormats.includes(rendition.instructions.fmt)) {
+        if (!supportedFormats.includes(rendition.instructions.fmt)) {
             // ...and tell the worker to generate a different intermediate format it can do
             // find first format that is supported (worker defines priority order)
-            const newFormat = supportedRenditionFormats.find(fmt => SUPPORTED_INPUT_FMT.includes(fmt));
+            const newFormat = supportedFormats.find(fmt => SUPPORTED_INPUT_FMT.includes(fmt));
             if (!newFormat) {
-                throw new GenericError(`Worker does not support generating a compatible image format. Requested: ${SUPPORTED_INPUT_FMT} Supported by worker: ${supportedRenditionFormats}`);
+                throw new GenericError(`Worker does not support generating a compatible image format. Requested: ${SUPPORTED_INPUT_FMT} Supported by worker: ${supportedFormats}`);
             }
 
             console.log(`post-processing: adjusted fmt for worker callback: '${rendition.instructions.fmt}' -> '${newFormat}'`);