--- a/lib/postprocessing/image.js
+++ b/lib/postprocessing/image.js
@@ -36,10 +36,10 @@ const SUPPORTED_OUTPUT_FMT = {
 // Note: order is important and defines priority for intermediate format - first entry will be most preferred
 const SUPPORTED_INPUT_TYPES = {
     "TIFF": "tif",  // most preferred format as it allows to preserve many features
-    "PNG": "png",   // lossless over
-    "JPEG": "jpeg", // ..lossy
-    "GIF": "gif",   // (should never get here in practice)
-    "BMP": "bmp"
+    "PNG":  "png",  // lossless over
+    "JPEG": "jpg",  // ..lossy
+    "GIF":  "gif",  // (should never get here in practice)
+    "BMP":  "bmp"
 };
 const SUPPORTED_INPUT_FORMAT_IMAGEMAGICK = Object.keys(SUPPORTED_INPUT_TYPES);
 const SUPPORTED_INPUT_FMT = Object.values(SUPPORTED_INPUT_TYPES);