--- a/lib/postprocessing/image.js
+++ b/lib/postprocessing/image.js
@@ -17,7 +17,9 @@ const { Dimensions } = require("./math");
 const util = require('util');
 const exec = util.promisify(require('child_process').exec);
 const { getAsset } = require('../storage');
-const { GenericError } = require('@adobe/asset-compute-commons');
+const { OpenwhiskActionName, GenericError } = require('@adobe/asset-compute-commons');
+
+const actionName = new OpenwhiskActionName().name;
 
 const WATERMARK_NAME = 'watermark.png';
 // list of supported output formats
@@ -200,7 +202,7 @@ async function prepareImagePostProcess(rendition, options) {
 
         const supportedFormats =
             options.supportedRenditionFormats
-            || WORKER_SUPPORTED_FMT[process.env.__OW_ACTION_NAME]
+            || WORKER_SUPPORTED_FMT[actionName]
             || DEFAULT_WORKER_SUPPORTED_FMT;
 
         // in case the worker cannot generate the final rendition format itself, we have to do it via post processing