--- a/lib/postprocessing/image.js
+++ b/lib/postprocessing/image.js
@@ -19,8 +19,6 @@ const exec = util.promisify(require('child_process').exec);
 const { getAsset } = require('../storage');
 const { OpenwhiskActionName, GenericError } = require('@adobe/asset-compute-commons');
 
-const actionName = new OpenwhiskActionName().name;
-
 const WATERMARK_NAME = 'watermark.png';
 // list of supported output formats
 // maps our fmt ('png') to imagemagick format ('PNG') https://imagemagick.org/script/formats.php
@@ -202,7 +200,7 @@ async function prepareImagePostProcess(rendition, options) {
 
         const supportedFormats =
             options.supportedRenditionFormats
-            || WORKER_SUPPORTED_FMT[actionName]
+            || WORKER_SUPPORTED_FMT[new OpenwhiskActionName().name]
             || DEFAULT_WORKER_SUPPORTED_FMT;
 
         // in case the worker cannot generate the final rendition format itself, we have to do it via post processing