--- a/lib/postprocessing/image.js
+++ b/lib/postprocessing/image.js
@@ -44,12 +44,12 @@ const SUPPORTED_INPUT_TYPES = {
 const SUPPORTED_INPUT_FORMAT_IMAGEMAGICK = Object.keys(SUPPORTED_INPUT_TYPES);
 const SUPPORTED_INPUT_FMT = Object.values(SUPPORTED_INPUT_TYPES);
 
-const DEFAULT_WORKER_SUPPORTED_FMT = [
+const DEFAULT_WORKER_SUPPORTED_OUTPUT_FMT = [
     "png"
 ];
 
 // NOTE: this is a temporary solution until workers have been updated to set options.supportedRenditionFormats
-const WORKER_SUPPORTED_FMT = {
+const WORKER_SUPPORTED_OUTPUT_FMT = {
     // https://git.corp.adobe.com/nui/worker-flite/blob/0e6dbe1e68ad9f226658bea3f7ba74c27a70073f/action/worker.js#L176
     // https://git.corp.adobe.com/nui/core/blob/5e1dc52e950a38f93599f1095b07ae163c40dac2/lib/workers.js#L225
     "worker-flite":     ["png", "jpg", "jpeg", "gif", "tif", "tiff"],
@@ -200,8 +200,8 @@ async function prepareImagePostProcess(rendition, options) {
 
         const supportedFormats =
             options.supportedRenditionFormats
-            || WORKER_SUPPORTED_FMT[new OpenwhiskActionName().name]
-            || DEFAULT_WORKER_SUPPORTED_FMT;
+            || WORKER_SUPPORTED_OUTPUT_FMT[new OpenwhiskActionName().name]
+            || DEFAULT_WORKER_SUPPORTED_OUTPUT_FMT;
 
         // in case the worker cannot generate the final rendition format itself, we have to do it via post processing
         if (!supportedFormats.includes(rendition.instructions.fmt)) {