--- a/lib/postprocessing/image.js
+++ b/lib/postprocessing/image.js
@@ -229,7 +229,6 @@ async function prepareImagePostProcess(rendition, options) {
 async function needsImagePostProcess(rendition) {
     const instructions = rendition.instructions;
 
-    // NOTE: already checked in prepareImagePostProcess() above, no need to check again?
     // verify if the requested format is one we can output from post-processing
     if (!isSupportedOutputFormat(instructions)) {
         console.log(`skipping post-processing because output fmt '${instructions.fmt}' is not supported`);
@@ -438,9 +437,9 @@ async function write(img, outfile) {
 
 /**
  * Runs image post processing on the (intermediate) rendition, to create the final rendition.
- * @param {*} intermediateRendition existing intermediate rendition created by the worker callback
- * @param {*} rendition final rendition to write to, including instructions
- * @param {*} directories working directories of the current activation
+ * @param {Rendition} intermediateRendition existing intermediate rendition created by the worker callback
+ * @param {Rendition} rendition final rendition to write to, including instructions
+ * @param {Object} directories working directories of the current activation
  */
 async function imagePostProcess(intermediateRendition, rendition, directories) {
     const img = await read(intermediateRendition.path);