--- a/lib/postprocessing/image.js
+++ b/lib/postprocessing/image.js
@@ -12,12 +12,14 @@
 
 'use strict';
 
-const gm = require("./gm-promisify");
-const { Dimensions } = require("./math");
+const { OpenwhiskActionName, GenericError, RenditionFormatUnsupportedError } = require('@adobe/asset-compute-commons');
+
 const util = require('util');
 const exec = util.promisify(require('child_process').exec);
+
 const { getAsset } = require('../storage');
-const { OpenwhiskActionName, GenericError } = require('@adobe/asset-compute-commons');
+const { Dimensions } = require("./math");
+const gm = require("./gm-promisify");
 
 const WATERMARK_NAME = 'watermark.png';
 // list of supported output formats
@@ -139,7 +141,7 @@ async function getImageMetadata(file) {
             return json.image;
         }
     } catch (e) {
-        throw new Error(`Reading metadata from intermediate rendition failed: ${e.message}`);
+        throw new GenericError(`Reading metadata from intermediate rendition failed: ${e.message}`);
     }
 }
 
@@ -196,7 +198,9 @@ function hasJpegQuality(metadata, instructions) {
  * or undefined if no changes are required.
  */
 async function prepareImagePostProcess(rendition, options) {
-    if (rendition.instructions.fmt && isSupportedOutputFormat(rendition.instructions)) {
+    const instructions = rendition.instructions;
+
+    if (instructions.fmt && isSupportedOutputFormat(instructions)) {
 
         const supportedFormats =
             options.supportedRenditionFormats
@@ -204,33 +208,48 @@ async function prepareImagePostProcess(rendition, options) {
             || DEFAULT_WORKER_SUPPORTED_OUTPUT_FMT;
 
         // in case the worker cannot generate the final rendition format itself, we have to do it via post processing
-        if (!supportedFormats.includes(rendition.instructions.fmt)) {
+        if (!supportedFormats.includes(instructions.fmt)) {
             // ...and tell the worker to generate a different intermediate format it can do
             // find first format that is supported (worker defines priority order)
             const newFormat = supportedFormats.find(fmt => SUPPORTED_INPUT_FMT.includes(fmt));
             if (!newFormat) {
-                throw new GenericError(`Worker does not support generating a compatible image format. Requested: ${SUPPORTED_INPUT_FMT} Supported by worker: ${supportedFormats}`);
+                throw new GenericError(
+                    "Worker does not support generating a compatible image format. " +
+                    `Requested: ${SUPPORTED_INPUT_FMT} Supported by worker: ${supportedFormats}`,
+                    "prepare_image_post_process"
+                );
             }
 
-            console.log(`post-processing: adjusted fmt for worker callback: '${rendition.instructions.fmt}' -> '${newFormat}'`);
+            console.log(`post-processing: adjusted fmt for worker callback: '${instructions.fmt}' -> '${newFormat}'`);
 
             rendition._forceImagePostProcess = true;
 
-            return Object.assign({}, rendition.instructions, { fmt: newFormat});
+            return Object.assign({}, instructions, { fmt: newFormat});
         }
     }
 
     // return undefined if no instruction change needed
 }
 
+function extension(filename) {
+    return filename ? filename.split(".").pop() : "";
+}
+
 /**
  * Returns true if the (intermediate) rendition needs image post processing.
  */
-async function needsImagePostProcess(rendition) {
+async function needsImagePostProcess(rendition, source) {
     const instructions = rendition.instructions;
 
     // verify if the requested format is one we can output from post-processing
     if (!isSupportedOutputFormat(instructions)) {
+        // if worker skipped AND we can't handle the output format, it effectively means
+        // the conversion is not supported and an error should be sent to the client
+        // normally this should not happen with proper worker routing
+        if (rendition.postProcess && rendition.postProcess.skippedProcessing) {
+            throw new RenditionFormatUnsupportedError(`Format ${instructions.fmt} is not supported for ${extension(source.path)} files`);
+        }
+
         console.log(`skipping post-processing because output fmt '${instructions.fmt}' is not supported`);
         return false;
     }
@@ -310,7 +329,8 @@ async function applyWatemark(img, intermediateRendition, rendition, directories)
     const watermarkSource = instructions.watermark.image;
     // download watermark
     if(!watermarkSource) {
-        throw new GenericError(`post-processing: Missing watermark.image in instructions : ${JSON.stringify(instructions)}`);
+        // TODO: needs a new client error InvalidRenditionInstructions or the like
+        throw new GenericError(`Missing watermark.image in instructions : ${JSON.stringify(instructions)}`, "image_post_process");
     }
     const asset = await getAsset(watermarkSource, directories.in, WATERMARK_NAME);
     // TODO: infuture should we enforce that watermark is a png ?