--- a/lib/prepare.js
+++ b/lib/prepare.js
@@ -17,7 +17,7 @@ const fse = require('fs-extra');
 
 const IN_DIRECTORY = "in";
 const OUT_DIRECTORY = "out";
-const POST_PROC_DIRECTORY = "post";
+const POST_PROCESSING_DIRECTORY = "post";
 const WORK_DIRECTORY = "work";
 
 async function createDirectories() {
@@ -40,18 +40,19 @@ async function createDirectories() {
 
     directories.in   = path.resolve(directories.base, IN_DIRECTORY);
     directories.out  = path.resolve(directories.base, OUT_DIRECTORY);
-    directories.post = path.resolve(directories.out, POST_PROC_DIRECTORY);
+    directories.postprocessing = path.resolve(directories.out, POST_PROCESSING_DIRECTORY);
 
     console.log(`work directory        : ${directories.base}`);
     console.log(`- source directory    : ${directories.in}`);
     console.log(`- renditions directory: ${directories.out}`);
+    console.log(`- post-processing dir : ${directories.postprocessing}`);
 
     // clean work directory if it might exist already
     await fse.remove(directories.base);
 
     await fse.mkdirs(directories.in);
     await fse.mkdirs(directories.out);
-    await fse.mkdirs(directories.post);
+    await fse.mkdirs(directories.postprocessing);
 
     // for test-worker framework, input and output are mounted at /in and /out
     // random access reading and writing from that mount can be problematic on Docker for Mac at least,