--- a/lib/prepare.js
+++ b/lib/prepare.js
@@ -97,7 +97,7 @@ async function cleanupDirectories(directories) {
         try {
             await fse.remove(directories.base);
         } catch (err) {
-            console.error(`Error while cleaning up work directories: ${err.message || err}`);
+            console.log(`Error while cleaning up work directories: ${err.message || err}`);
             return false;
         }
     }