--- a/lib/rendition.js
+++ b/lib/rendition.js
@@ -253,7 +253,11 @@ class Rendition {
      * Redacts URL properties.
      */
     instructionsForEvent() {
-        const obj = { ...this.originalInstructions };
+        return Rendition.redactInstructions(this.originalInstructions);
+    }
+
+    static redactInstructions(instructions) {
+        const obj = { ...instructions };
         // remove target URLs, could be sensitive
         delete obj.target;
         return obj;