--- a/lib/rendition.js
+++ b/lib/rendition.js
@@ -260,6 +260,7 @@ class Rendition {
         const obj = { ...instructions };
         // remove target URLs, could be sensitive
         delete obj.target;
+        delete obj.userData;
         return obj;
     }