--- a/lib/rendition.js
+++ b/lib/rendition.js
@@ -260,7 +260,6 @@ class Rendition {
         const obj = { ...instructions };
         // remove target URLs, could be sensitive
         delete obj.target;
-        delete obj.userData;
         return obj;
     }