--- a/lib/rendition.js
+++ b/lib/rendition.js
@@ -242,6 +242,8 @@ class Rendition {
     changeInstructions(newInstructions) {
         this._originalInstructions = this.instructions;
         this.instructions = newInstructions;
+
+        Object.freeze(this._originalInstructions);
     }
 
     get originalInstructions() {