--- a/lib/storage.js
+++ b/lib/storage.js
@@ -114,7 +114,7 @@ async function putRendition(rendition, directories) {
     // Note: validation has happened in validate.js before
 
     if (process.env.WORKER_TEST_MODE) {
-        if (rendition.directory === directories.post) {
+        if (rendition.directory === directories.postprocessing) {
             // move the post processing rendition to the rendition path as expected by "aio asset-compute test-worker"
             const newRendition = new Rendition(rendition.instructions, directories.out, rendition.index);