--- a/lib/utils/timer.js
+++ b/lib/utils/timer.js
@@ -40,7 +40,7 @@ class Timer {
                 this._current = hrtimeToSeconds(process.hrtime(this._id));
                 this._total += this._current;
             } catch (e) {
-                console.error("error getting timing metrics:", e.message || e);
+                console.log("error getting timing metrics:", e.message || e);
             }
 
             delete this._id;