--- a/lib/webaction.js
+++ b/lib/webaction.js
@@ -107,7 +107,7 @@ async function invokeSelfAsync(params) {
 
     } catch (e) {
         const msg = `Async invocation of ${actionName} failed with HTTP status code ${e.statusCode}`;
-        console.error(msg, e);
+        console.log(msg, e);
         const code = e.statusCode === 429 ? 429 : 500;
         throw httpError(code, msg);
     }