… lib/shell/shellscript.js
… lib/utils/type.js
… lib/utils/utils.js
… lib/worker.js
… test-worker/package.json
… test-worker/test/asset-compute/worker/skipped-unsupported-output/file.png
… test-worker/test/asset-compute/worker/skipped-unsupported-output/params.json
… test-worker/worker.js
… test/api.test.js
… test/postprocess.test.js
… test/rendition.content.test.js
… test/rendition.test.js
… test/storage.test.js
… test/type.test.js