New Feature - For [NUI-867](https://jira.corp.adobe.com/browse/NUI-867). Required for webp (or any other new output format) support: #66.

### Overview (read this)

This turned out to be a slightly larger change 😉 The change is actually a key part of the post processing logic, but we were lucky that so far it worked without it. However, with `webp` as new format that none of the worker (tools) support, it is required. The problem is

1. if a rendition asks for `webp`
2. but a worker callback/tool can only generate `png`
3. we must not blindly pass `fmt=webp` to the worker callback but `fmt=png` instead
4. and post processing would then convert it the png to the final `webp`

Otherwise the worker callbacks would see `fmt=webp` in step 3 and bail out with errors.

This [diagram](https://git.corp.adobe.com/nui/nui/raw/master/images/asset-compute-sdk-post-processing.png) illustrates the overall flow, with the "Post-processing decision (1)" box (top left) being the new element added in this PR, implemented by [`prepareImagePostProcess()`](https://github.com/adobe/asset-compute-sdk/pull/65/files#diff-c8cfb029f1e204a8f6da461ffa1464b0R200).

This lead to a few follow up changes around the sdk, see change list below. If there are any questions, happy to go through in a short mob session.

Along the way I improved a few things mostly around logging, and some test structure improvement. Let me know if I should separate them out.

### Changes
- (main logic) update `fmt` in renditions instructions for worker callback to a different intermediate format such as `png` if needed
  - logic only kicks in if the `fmt` field is set AND it is an output format supported by image post processing, otherwise it will leave instructions untouched
  - custom workers are not affected unless they use the `fmt` field (which we should probably reserve...)
- (**worker api**) workers can now set `options.supportedRenditionFormats`
   - until that is provided by all workers, the code includes a _temporary_ [list of supported renditions for our current workers](https://github.com/adobe/asset-compute-sdk/pull/65/files#diff-c8cfb029f1e204a8f6da461ffa1464b0R53-R93)
   - this allows dependent module & integration tests to pass w/o changes to any worker
   - there is not one list that works for all. for example, `cameraraw` can only generate `jpg + tiff`, while `pdfrasterizer` only supports `png` output
  - otherwise defaults to `["png"]`
- (impl) new [preparePostProcess()](https://github.com/adobe/asset-compute-sdk/pull/65/files#diff-644d8ede25bb88427ccf4229d03bba1bR312) in `worker.js` calling a new [prepareImagePostProcess()](https://github.com/adobe/asset-compute-sdk/pull/65/files#diff-c8cfb029f1e204a8f6da461ffa1464b0R200) in `image.js` function making the new decision
  - `prepareImagePostProcess()` runs _before_ the callback
  - `needsImagePostProcess()` runs _after_ the callback and decides if the generated rendition needs post processing
  - `imagePostProcess()` called to do post processing if needed
  - there should be the opportunity to extend this for other future post processing logic (e.g. sensei) where we plugin other parts into the `Worker.preparePostProcess()` (and co) functions
- (tests) `test-worker/worker.js` simulates a worker only supporting some output formats and fails if incorrect format is requested
- (corollary) change Rendition class to support generating the path dynamically, in case `instructions.fmt` changed
  - better rendition file separation because of that dynamic path change
  - with post processing writing to `post/` directory instead of "post-" filename prefix
  - see example below
- (corollary) throw error if options file from shell script cannot be parsed instead of silently ignoring it, as in case of post processing this could have lead to returning an incorrect rendition (the intermediate one, w/o the required post processing)

Further improvements but unrelated:

- moved content type detection from rendition.js into separate module type.js to make it easier testable
- adjusted rendition tests accordingly and using mockRequire to simulate 'file' failures
- improved logging
   - now prefixing per rendition logs with `[rendition 0]` and the right index
   - logging asset download and WORKER_TEST_MODE file moving/copying
   - see example below
- also made one slow api test faster (5 sec to 600ms)

---

### Examples

Example of various rendition files (from log below):

* source: `in/file.gif`
* rendition generated by worker callback: `out/rendition0.png`
* rendition generated by post processing: `out/post/rendition0.gif`
* (WORKER_TEST_MODE: copied back into `out` folder: `out/rendition0.gif` (so that test-worker finds it))

Logging example:

```
2020-09-29T05:53:15.718556500Z worker worker Test: gif-converttodpi -----------------------------------------------
2020-09-29T05:53:15.718578200Z work directory        : /nodejsAction/KbNvAqfg/work/kfnjogzl
2020-09-29T05:53:15.718598400Z - source directory    : /nodejsAction/KbNvAqfg/work/kfnjogzl/in
2020-09-29T05:53:15.718618400Z - renditions directory: /nodejsAction/KbNvAqfg/work/kfnjogzl/out
2020-09-29T05:53:15.751385400Z WORKER_TEST_MODE: asset path: /nodejsAction/KbNvAqfg/work/kfnjogzl/in/file.gif
2020-09-29T05:53:15.751778900Z source downloaded in 0.000 seconds
2020-09-29T05:53:15.752455900Z [rendition 0] post-processing: adjusted fmt for worker callback: 'gif' -> 'png'
2020-09-29T05:53:15.753182400Z [rendition 0] original instructions: { convertToDpi: { xdpi: 48, ydpi: 48 }, fmt: 'gif' }
2020-09-29T05:53:15.753604000Z [rendition 0] generating rendition0.png...
2020-09-29T05:53:15.754057400Z [rendition 0] instructions for 'worker()' callback: { convertToDpi: { xdpi: 48, ydpi: 48 }, fmt: 'gif' }
2020-09-29T05:53:15.754482400Z [rendition 0] [test worker] copying source to rendition: /nodejsAction/KbNvAqfg/work/kfnjogzl/in/file.gif to /nodejsAction/KbNvAqfg/work/kfnjogzl/out/rendition0.png
2020-09-29T05:53:15.755380100Z [rendition 0] worker() callback generated rendition in 0.001 seconds: rendition0.png
2020-09-29T05:53:15.768133500Z [rendition 0] .....................................................................................................
2020-09-29T05:53:15.768571900Z [rendition 0] Version: ImageMagick 7.0.9-19 Q16 x86_64 2020-01-26 https://imagemagick.org
2020-09-29T05:53:15.768600600Z Copyright: © 1999-2020 ImageMagick Studio LLC
2020-09-29T05:53:15.768621900Z License: https://imagemagick.org/script/license.php
2020-09-29T05:53:15.768643300Z Features: Cipher DPC HDRI Modules OpenMP(3.1) 
2020-09-29T05:53:15.768664200Z Delegates (built-in): bzlib cairo djvu fftw fontconfig freetype gslib jng jp2 jpeg lcms ltdl lzma openexr pangocairo png ps raqm rsvg tiff webp wmf x xml zlib
2020-09-29T05:53:15.769027100Z [rendition 0] .....................................................................................................
2020-09-29T05:53:15.832995800Z [rendition 0] post-processing image rendition rendition0.png => post/rendition0.gif
2020-09-29T05:53:15.835456300Z [rendition 0] imagemagick command: convert /nodejsAction/KbNvAqfg/work/kfnjogzl/out/rendition0.png[0] -auto-orient -orient Undefined -colorspace sRGB -units pixelsperinch -resample 48x48 -set comment 'Generated by Adobe Asset Compute SDK post-processing.' /nodejsAction/KbNvAqfg/work/kfnjogzl/out/post/rendition0.gif
2020-09-29T05:53:16.088282600Z [rendition 0] post-processing rendition0.png finished successfully in 0.333 seconds
2020-09-29T05:53:16.089957300Z [rendition 0] WORKER_TEST_MODE: moving /nodejsAction/KbNvAqfg/work/kfnjogzl/out/post/rendition0.gif to /nodejsAction/KbNvAqfg/work/kfnjogzl/out/rendition0.gif
2020-09-29T05:53:16.093562700Z [rendition 0] event not sent: rendition_created {"rendition":{"convertToDpi":{"xdpi":48,"ydpi":48},"fmt":"gif"},"metadata":{},"type":"rendition_created","date":"2020-09-29T05:53:16.093Z","requestId":"Test: gif-converttodpi -----------------------------------------------","source":"file.gif"}
2020-09-29T05:53:16.127828700Z XXX_THE_END_OF_A_WHISK_ACTIVATION_XXX
```