@tmathern 2020-09-30
two things that come to my mind:
- We need to update the supported formats;
- We need to update the documentation about postprocessing and custom workers, to explain what you formulated in the PR description.