@alexkli 2020-09-30
@tmathern 
1. We need to update the supported formats
    - => tracked at https://jira.corp.adobe.com/browse/NUI-873
2. We need to update the documentation about postprocessing and custom workers, to explain what you formulated in the PR description.
    - => nothing should change, as post processing is not available for custom workers at this point (no imagemagick available). we will document it once we get there, but then it will certainly have evolved.