@jdelbick 2020-09-30
not going to lie, it was hard to review all these changes in a single PR @alexkli 

some more comments:
1. I could see the log changes in a separate PR
2. I could see the content-type changes in a separate PR (especially because you had to touch some unit tests with this change that add to the amount of material to review)
3. worker test cases...are these mostly covered in you supporting PR: https://github.com/adobe/asset-compute-sdk/pull/66?test cases like this (that would previously have failed):

> if a rendition asks for webp
but a worker callback/tool can only generate png
we must not blindly pass fmt=webp to the worker callback but fmt=png instead
and post processing would then convert it the png to the final webp