@alexkli 2020-09-30
@jdelbick 

1. yeah, I was hoping it's related enough 😉  this new feature (more complex decisions) required more and better logs after all... FWIW, initially I had this together with #66 ... then I split it up already so the pure webp change is isolated.
2. the content-type changes where triggered by `Rendition.path` being a generated property now (instead of being set in the constructor) due to the rendition file/dir separation. a bit tricky to separate them out.
3. worker test cases: this is mostly covered by [this change in the test worker](https://github.com/adobe/asset-compute-sdk/pull/65/files#diff-372d7ae33e7b72869ab310eeecd2ba86R30-R34) which makes a lot of the existing test cases fail without the functional change.