@alexkli 2020-10-09
More changes:

- for worker-flite: throw `RenditionFormatUnsupportedError` if worker callback skips processing (`rendition.postProcess.skippedProcessing == true`) and output format is not supported by post-processing
  - tested by new test case `skipped-unsupported-output`
- ensure a `GenericError` is thrown if no asset compute specific error is used
- only use `console.log()` so that the `[rendition N]` log prefix shows everywhere