@alexkli 2020-10-09
@tmathern @jdelbick Merging this for the next steps with webp (i.e. #66 and testing), but not releasing yet, so we can address any feedback on the additional changes added since you all reviewed https://github.com/adobe/asset-compute-sdk/pull/65/commits/24af07b9a54d834c73f61afbffc023440d64894d.